refactor: modularize tutorial
